+++ CHANGELOG.md
@@ -3,6 +3,13 @@
 The format is based on and uses the types of changes according to [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Changed
+
+- BREAKING CHANGE: Rename scripts folder to SPSWakeUP
+- BREAKING CHANGE: \workflows\release.yml update scripts to SPSWakeUP
+
 ## [3.0.2] - 2024-12-16
 
 ### Changed